# backend/contracts/alojamientos-openapi.yaml
...
openapi: 3.0.3
...
info:
  # El nombre de la API que aparecerá en la documentación
  title: GDS Alojamientos Core API
  # Una descripción detallada de lo que hace este microservicio
  description: Microservicio centralizado para búsqueda, disponibilidad, reservas (órdenes), postventa de alojamientos y mensajería con la propiedad.
    El dueño de la reserva (ownerId) se infiere del sub del token JWT. La lógica de pagos pertenece a otros dominios/APIs.
  # La versión actual del contrato/API
  version: 1.0.0
  # Información de contacto de los responsables de la API
  contact:
    name: Joselyn Cadena
    email: jlcadenac@puce.edu.ec

# Lista de servidores donde la API está alojada
servers:
# URL del entorno de Producción
- url: https://api.booking-hub.com/alojamientos/v1
  description: Producción
# URL del entorno de Pruebas o Sandbox
- url: https://sandbox.api.booking-hub.com/alojamientos/v1
  description: Sandbox

# Configuración global de seguridad. Todos los endpoints exigirán esto por defecto.
security:
# Especifica que usaremos el esquema de seguridad OAuth2 definido más abajo
- OAuth2Security: []

# Categorías visuales para organizar la documentación (Swagger)
...
paths:
...
  /search:
    # Método HTTP POST
    post:
      # Asignamos este endpoint a la categoría "Búsqueda y Catálogo"
      tags:
      - Búsqueda y Catálogo
      # Un resumen corto de lo que hace el endpoint
      summary: Búsqueda de alojamientos
      # Sobreescribe la seguridad global: este endpoint es público (no requiere token)
      security: []
      # Parámetros que la petición recibe en la URL o en los Headers
      parameters:
      # Definimos un header obligatorio llamado X-Device-Fingerprint
      - name: X-Device-Fingerprint
        in: header
        required: true
        schema:
          type: string
      # El cuerpo de la petición (JSON que el cliente debe enviar)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              # Hace referencia al esquema 'SearchAccommodationRequest' definido en components
              $ref: '#/components/schemas/SearchAccommodationRequest'
      # Las posibles respuestas que este endpoint puede devolver
      responses:
        # Si todo sale bien (Código 200 OK)
        '200':
          description: Alojamientos encontrados
          headers:
            Cache-Control:
              schema:
                type: string
                example: 'public, max-age=300'
            X-API-Deprecation-Date:
              schema:
                type: string
                format: date
                description: Fecha prevista de baja de esta versión de la API.
          content:
            application/json:
              schema:
                # Hace referencia al esquema de respuesta
                $ref: '#/components/schemas/SearchAccommodationResponse'
        # Si el cliente envía datos inválidos (Código 400)
        '400':
          $ref: '#/components/responses/ProblemDetails400'
        # Si el cliente hace demasiadas peticiones (Código 429)
        '429':
          $ref: '#/components/responses/ProblemDetails429'
  
  # Endpoint para ver disponibilidad y precios
...
  /details:
    post:
      tags:
      - Búsqueda y Catálogo
      summary: Obtener detalles extendidos de los alojamientos
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccommodationDetailsRequest'
      responses:
        '200':
          description: Detalles de los alojamientos solicitados
          headers:
            Cache-Control:
              schema:
                type: string
                example: 'public, max-age=300'
            X-API-Deprecation-Date:
              schema:
                type: string
                format: date
                description: Fecha prevista de baja de esta versión de la API.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccommodationDetailsResponse'
  
  # Endpoint para ver qué hoteles han modificado su información recientemente
...
components:
...
  securitySchemes:
    # Definimos cómo funciona el OAuth2
    OAuth2Security:
      # Tipo de seguridad
      type: oauth2
      # Flujos permitidos en OAuth2
      flows:
        # Flujo para usuarios que inician sesión (ej. con correo y contraseña)
        authorizationCode:
          authorizationUrl: https://auth.booking-hub.com/oauth2/authorize
          tokenUrl: https://auth.booking-hub.com/oauth2/token
          # Permisos disponibles (scopes)
          scopes:
            alojamientos:read: Leer información de alojamientos, disponibilidades y reservas
            alojamientos:book: Crear y alterar reservas
            alojamientos:cancel: Cancelar reservas
            alojamientos:webhooks: Gestionar webhooks
        # Flujo para sistemas automáticos o máquinas (backend a backend)
        clientCredentials:
          tokenUrl: https://auth.booking-hub.com/oauth2/token
          scopes:
            alojamientos:read: (B2B) Leer
            alojamientos:book: (B2B) Comprar
            alojamientos:cancel: (B2B) Cancelar
            alojamientos:webhooks: (B2B) Webhooks
  
  # Moldes de Respuestas de Error. Todos usan el esquema 'ProblemDetails'
  responses:
    ProblemDetails400:
      description: Petición inválida
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
    ProblemDetails429:
      description: Demasiadas peticiones
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
  
  # Moldes de Datos (Schemas). Definimos los JSON exactos que viajan en las peticiones.
  schemas:
...
    ProblemDetails:
      # Es un objeto JSON
      type: object
      # No permite campos extra que no estén definidos aquí
      additionalProperties: false
      # Estos campos son obligatorios
      required:
      - type
      - title
      - status
      - code
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        code:
          type: string
          # Valores posibles para el código de error
          enum:
          - VALIDATION_FAILED
          - ROOM_NO_LONGER_AVAILABLE
          - PRICE_CHANGED
          - BOOKING_NOT_CONFIRMED
          - CANCELLATION_NOT_ALLOWED
          - RATE_LIMIT_EXCEEDED
          - PAYMENT_REFERENCE_INVALID
          - PAYMENT_NOT_AUTHORIZED
        # Lista de parámetros inválidos para explicar exactamente qué falló
        invalidParams:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              reason:
                type: string
    
    # Esquema que representa a la persona que hace la reserva
    Booker:
      type: object
      required:
      - country
      - platform
      properties:
        country:
          type: string
          # Regresión regular: obliga a que sean exactamente 2 letras minúsculas (Ej. 'us', 'mx')
          pattern: ^[a-z]{2}$
          description: Código ISO 3166-1 alpha-2 del país del comprador.
        platform:
          type: string
          enum:
          - android
          - desktop
          - ios
          - mobile
          - tablet
        state:
          type: string
          pattern: ^[a-z]{2}$
        travel_purpose:
          type: string
          enum:
          - business
          - leisure
        user_groups:
          type: array
          items:
            type: string
            enum:
            - authenticated
    
    # Esquema que representa a los huéspedes
    AccommodationsGuests:
      type: object
      required:
      - number_of_adults
      - number_of_rooms
      properties:
        number_of_adults:
          type: integer
          # Mínimo 1 adulto
          minimum: 1
        number_of_rooms:
          type: integer
          minimum: 1
        children:
          type: array
          items:
            type: integer
            description: Edades de los niños
        allocation:
          type: array
          items:
            type: object
            properties:
              adults:
                type: integer
              children:
                type: array
                items:
                  type: integer
    
    # JSON que se envía al buscar alojamientos
    SearchAccommodationRequest:
      type: object
      required:
      - booker
      - checkin
      - checkout
      - guests
      properties:
        # Inyecta el esquema de Booker aquí
        booker:
          $ref: '#/components/schemas/Booker'
        checkin:
          type: string
          # Formato de fecha estándar (YYYY-MM-DD)
          format: date
        checkout:
          type: string
          format: date
        city:
          type: integer
        country:
          type: string
          pattern: ^[a-z]{2}$
        guests:
          $ref: '#/components/schemas/AccommodationsGuests'
        extras:
          type: array
          items:
            type: string
            enum:
            - extra_charges
            - products
        currency:
          type: string
          # Exactamente 3 letras mayúsculas para la moneda (Ej. USD, EUR)
          pattern: ^[A-Z]{3}$
        rows:
          type: integer
          minimum: 10
          maximum: 100
          default: 100
        page:
          type: string
    
    # Respuesta de una búsqueda
    SearchAccommodationResponse:
      type: object
      properties:
        request_id:
          type: string
        data:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              url:
                type: string
        next_page:
          type: string
          # Puede ser nulo si no hay página siguiente
          nullable: true
    
    # JSON enviado para verificar disponibilidad
...
    AccommodationDetailsRequest:
      type: object
      properties:
        accommodations:
          type: array
          items:
            type: integer
        city:
          type: integer
        country:
          type: string
        extras:
          type: array
          items:
            type: string
            enum:
            - description
            - bundles
            - facilities
            - payment
            - photos
            - policies
            - rooms
        languages:
          type: array
          items:
            type: string
    AccommodationDetailsResponse:
      type: object
      properties:
        request_id:
          type: string
        data:
          type: array
          items:
            type: object
        next_page:
          type: string
          nullable: true
...